Changes for domain getters
