+++ packages/adapters/txservice/src/shared/provider.ts
@@ -1,4 +1,4 @@
-import { delay, getChainIdFromDomain, parseHostname } from "@connext/nxtp-utils";
+import { chainIdToDomain, delay, domainToChainId, parseHostname } from "@connext/nxtp-utils";
 import { providers, utils } from "ethers";
 
 import { parseError, RpcError, ServerError, StallTimeout } from "./errors";
@@ -58,7 +58,7 @@ export class SyncProvider extends StaticJsonRpcProvider {
   ) {
     // NOTE: super (StaticJsonRpc) uses the hard-coded chainId when instantiated for all future
     // .getNetwork() requests, so it is important to use the chainId here, not the domain
-    super(_connectionInfo, getChainIdFromDomain(domain));
+    super(_connectionInfo, domainToChainId(domain));
     this.connectionInfo = typeof _connectionInfo === "string" ? { url: _connectionInfo } : _connectionInfo;
     this.name = parseHostname(this.connectionInfo.url)
       ? parseHostname(this.connectionInfo.url)!.split(".").slice(0, -1).join(".")
