+++ packages/agents/watcher/src/watcher.ts
@@ -10,9 +10,9 @@ import {
   getChainData,
   Logger,
   RequestContext,
+  domainToChainId,
 } from "@connext/nxtp-utils";
 import { utils, Wallet } from "ethers";
-import { domainToChainId } from "@connext/smart-contracts";
 
 import { bindServer, bindInterval } from "./bindings";
 import { getConfig } from "./config";
