+++ packages/deployments/contracts/deploy/02_deployConnext.ts
@@ -6,8 +6,8 @@ import { DeploymentSubmission } from "hardhat-deploy/dist/types";
 import { SKIP_SETUP } from "../src/constants";
 import { getConnectorName, getDeploymentName, getProtocolNetwork, getRelayerProxyConfig } from "../src/utils";
 import { FacetOptions, getProposedFacetCuts, getUpgradedAbi } from "../deployHelpers";
-import { chainIdToDomain } from "../src";
 import { MESSAGING_PROTOCOL_CONFIGS } from "../deployConfig/shared";
+import { chainIdToDomain } from "@connext/nxtp-utils";
 
 /**
  * Hardhat task defining the contract deployments for Connext
