+++ packages/deployments/contracts/tasks/preflight.ts
@@ -1,9 +1,10 @@
 import { BigNumber, constants, Contract, utils } from "ethers";
 import { isAddress } from "ethers/lib/utils";
 import { task } from "hardhat/config";
+import { chainIdToDomain } from "@connext/nxtp-utils";
 
 import { Env, getDeploymentName, mustGetEnv } from "../src/utils";
-import { canonizeId, chainIdToDomain } from "../src";
+import { canonizeId } from "../src";
 
 // Default amount of tokens to mint / add liquidity for.
 const DEFAULT_AMOUNT = "2500000";
