+++ packages/utils/package.json
@@ -52,10 +52,10 @@
     "nyc": "15.1.0",
     "rimraf": "3.0.2",
     "sinon": "15.0.1",
+    "sinon-chai": "3.7.0",
     "ts-mocha": "10.0.0",
     "tslib": "2.5.0",
-    "typescript": "4.9.5",
-    "sinon-chai": "3.7.0"
+    "typescript": "4.9.5"
   },
   "publishConfig": {
     "registry": "https://registry.npmjs.org",
